feat(vendors): Allow adding ingredients directly from vendor view
This change introduces the ability to add a new ingredient directly from a vendor's detail view. When clicking an "Add Ingredient" button on a vendor card, the user is navigated to the "Add Ingredient" form, with the corresponding vendor pre-selected. This streamlines the workflow for adding ingredients associated with a specific supplier.

The `VendorsView` now accepts an `onAddIngredientForVendor` prop, and `VendorCard` has been updated with an `onAddIngredient` handler. The `RecipeCosting` component also gains a `preselectedVendorId` state to manage this new flow, and `AddIngredientView` is updated to accept and utilize this preselected vendor ID.